Added WebDav, updated Readme
