+++ README.md
@@ -1,9 +1,32 @@
+<p align="center">
+    <img src="https://i.imgur.com/h8wLVqH.png" style="width: 69%">
+</p>
+
 # ServerBackup
 
-SourceCode of version 'v2.9.6'
+ServerBackup is a Minecraft mod that enables server administrators to create backups of worlds or the entire server. It provides convenient functionality for safeguarding your Minecraft server's data.
+Features
 
-Website: https://server-backup.net/
+* Create backups of individual worlds or the entire server
+* Customizable backup schedules and intervals
+* Multithreaded backup process for improved performance
+* Easy-to-use configuration options
 
-Spigot: https://www.spigotmc.org/resources/world-server-backup-1-8-1-18-x-multithreaded.79320/
+## Usage
 
-Discord server for support or suggestions: https://discord.gg/rNzngsCWFC
+* Install the ServerBackup mod on your Minecraft server.
+* Configure the backup settings in the mod's configuration file.
+* Run the server to automatically create backups based on your specified schedule.
+
+You can also check the [usage guide](https://github.com/P529/ServerBackup/wiki/Usage-guide)
+
+## Contributing
+
+Contributions, bug reports, and feature requests are welcome. Please submit them through the GitHub repository or join our Discord server for support and suggestions.
+Links
+
+* Website: [ServerBackup Website](https://server-backup.net/)
+* Spigot page: [ServerBackup on Spigot](https://www.spigotmc.org/resources/server-backup-ingame-dropbox-ftp-backup-1-8-1-19-multithreaded.79320/)
+* Discord server: [ServerBackup Discord](https://discord.gg/rNzngsCWFC)
+
+If you have any questions or need further assistance, please feel free to reach out to us on Discord or through the GitHub repository.
