+++ src/main/java/de/seblii/serverbackup/commands/ServerCommand.java
@@ -1,0 +1,8 @@
+package de.seblii.serverbackup.commands;
+
+import de.seblii.serverbackup.ServerBackup;
+import org.bukkit.command.CommandSender;
+
+public interface ServerCommand {
+    boolean execute(CommandSender sender, String[] args, ServerBackup backup);
+}
